File Handling update 5
This update allows the inclusion of other tables inside of the table. It does this by creating a file reference and then creating a new file for that new table. Then, during the reformatting, it reads the data and adds it back to the overall table. Any potential bugs yet to be determiuned