File System Update
The current reformatter has not been designed to detect formatted data in a given file that is outside of segments. As such, it cannot find any "(" and so glosses over the file entirely. This needs to be fixed.  An addition txt file has been added of a later version which works for any kind of table however is also suffering from the same issue. This is the next step, as the function to isolate each segment needs entirely re-doing to allow the inclusion of variables outside of segments.